Add script to automate updating device list
Don't want to install VSCode just for that feature of the Connect IQ
plugin.
So painstakingly asked (many times...) ChatGPT to write a script to use
the list of devices in the `Devices` folder.
New devices are by default supported. Then I test them out and comment
out in the `manifest.xml` file if I deem them not possible to support
(very few devices, but still, it happens). This also helps me keep
better track of which devices I've tested in the past and did not add
support for them (usually because they don't have a screen or some weird
stuff like that).
Bash Gods, please excuse the horrible script, it's probably
unnecessarily convoluted. And only works on MacOS, most likely...