Handle full redraws better while avoiding unnecessary clear ops
So, the previous optimization was a bit too naive: the screen might be
requested to updated by the system. For example, when a notification is
shown, when the user brings up the quick menu, even when they bring up
the in-app menu.
So instead, let's by "default" clear the screen. But when we know that
the request to update the view is coming from the timers, then we
_should_ be safe to update only part of the screen.